bot/requirements.txt
aiogram==3.13.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
asyncpg==0.29.0
redis==5.0.8
python-dotenv==1.0.1
pydantic==2.8.2
aiofiles==23.2.1
python-multipart==0.0.9